delete printf and add uart msg

+++ source/sensors.h
@@ -90,7 +90,7 @@ typedef struct
 }break_data_t;
 
 typedef struct{
-	int16_t direction_value;
+	float direction_value;
 	uint16_t calibration_direction_value;
 } direction_data_t;
